can start deferred chain without F now. Some tests to fix multi
